Move incomeStatement types under lib-common
