+++ frontend/src/citizen-frontend/income-statements/IncomeStatements.tsx
@@ -6,7 +6,7 @@ import Footer from '../Footer'
 import { useTranslation } from '../localization'
 import AddButton from 'lib-components/atoms/buttons/AddButton'
 import { useHistory } from 'react-router-dom'
-import { IncomeStatement } from './types/income-statement'
+import { IncomeStatement } from 'lib-common/api-types/incomeStatement'
 import { Loading, Result } from 'lib-common/api'
 import { getIncomeStatements } from './api'
 import { SpinnerSegment } from 'lib-components/atoms/state/Spinner'
