+++ frontend/src/citizen-frontend/income-statements/types/body.ts
@@ -1,7 +1,7 @@
 import LocalDate from 'lib-common/local-date'
 import { stringToInt } from 'lib-common/utils/number'
 import { UUID } from 'lib-common/types'
-import { HighestFee, Income } from './income-statement'
+import { HighestFee, Income } from 'lib-common/api-types/incomeStatement'
 import * as Form from './form'
 
 export type HighestFeeBody = Omit<HighestFee, 'id'>
