+++ frontend/src/citizen-frontend/income-statements/types/common.ts
@@ -1,33 +1,4 @@
-export type IncomeSource = 'INCOMES_REGISTER' | 'ATTACHMENTS'
-
-export const otherIncome = [
-  'PENSION',
-  'ADULT_EDUCATION_ALLOWANCE',
-  'SICKNESS_ALLOWANCE',
-  'PARENTAL_ALLOWANCE',
-  'HOME_CARE_ALLOWANCE',
-  'FLEXIBLE_AND_PARTIAL_HOME_CARE_ALLOWANCE',
-  'ALIMONY',
-  'INTEREST_AND_INVESTMENT_INCOME',
-  'RENTAL_INCOME',
-  'UNEMPLOYMENT_ALLOWANCE',
-  'LABOUR_MARKET_SUBSIDY',
-  'ADJUSTED_DAILY_ALLOWANCE',
-  'JOB_ALTERNATION_COMPENSATION',
-  'REWARD_OR_BONUS',
-  'RELATIVE_CARE_SUPPORT',
-  'BASIC_INCOME',
-  'FOREST_INCOME',
-  'FAMILY_CARE_COMPENSATION',
-  'REHABILITATION',
-  'EDUCATION_ALLOWANCE',
-  'GRANT',
-  'APPRENTICESHIP_SALARY',
-  'ACCIDENT_INSURANCE_COMPENSATION',
-  'OTHER_INCOME'
-] as const
-
-export type OtherIncome = typeof otherIncome[number]
+import { otherIncome } from 'lib-common/api-types/incomeStatement'
 
 export const attachmentType = [
   ...otherIncome,
