+++ frontend/src/citizen-frontend/income-statements/types/form.ts
@@ -1,7 +1,7 @@
-import { IncomeSource, OtherIncome } from './common'
 import { Attachment } from 'lib-common/api-types/attachment'
-import * as IncomeStatement from './income-statement'
+import * as IncomeStatement from 'lib-common/api-types/incomeStatement'
 import LocalDate from 'lib-common/local-date'
+import { IncomeSource, OtherIncome } from 'lib-common/api-types/incomeStatement'
 
 export interface IncomeStatementForm {
   startDate: string
